refactor + extract TeamMembers component
